Added modularity to puzzle gameplay
